requirements.txt
absl-py==2.3.1
astunparse==1.6.3
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.11.2
keras-facenet==0.3.2
libclang==18.1.1
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
mtcnn==1.0.0
mysql-connector-python==9.4.0
namex==0.1.0
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==6.31.1
Pygments==2.19.2
requests==2.32.4
rich==14.1.0
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
